refactor discord webhook command
